smart-prompting-assistant/requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
flask-cors>=4.0.0
Jinja2>=3.1.2
python-dotenv>=1.0.0